+++ Back-End/HealthNest/src/main/java/com/healthnest/dto/UserDTO.java
@@ -1,16 +1,16 @@
 package com.healthnest.dto;
 
-import com.healthnest.dto.enums.Gender;
+import com.healthnest.model.enums.Gender;
 
 import lombok.Data;
 
 @Data
 public class UserDTO {
-	private Integer userId;
-	private String name;
-	private Gender gender;
-	private String password;
-	private String email;
-	private String dateOfBirth;
-	private String phoneNo;
+    private Long userId;  // Changed from Integer to Long
+    private String name;
+    private Gender gender;
+    private String password;
+    private String email;
+    private String dateOfBirth;
+    private String phoneNo;
 }
